Update the runtime version in bash script
The bash script to submit training job for pets detection has runtime-version of 1.8. This will trigger `TypeError: non_max_suppression() got an unexpected keyword argument 'score_threshold'` on the Google Cloud since 1.8 and older does not support this keyword argument. Therefore, update this runtime version to 1.9, which is the most recent runtime version that is published on June 27, 2018. 

See 
https://github.com/tensorflow/models/issues/5056
https://cloud.google.com/ml-engine/docs/tensorflow/runtime-version-list